Pick Chebyshev coefficients by sorting
Instead of truncating the tail of the vector containing Chebyshev
coefficients, the code now order them according to their absolute value
and save the ones with the highest value. This has required some API
reworkings, as now it is no longer enough to save the number of
Chebyshev coefficients: their positions are therefore kept using a
bitmask.