Remove a few memory leaks
One of the two memory leaks solved by this commit led to a change in the way
`pcomp_find_chebyshev_mask` works: previously it allocated a newly array of
`size_t` elements in the heap for each call (and forgot to free it). Now the
chunk is kept within `pcomp_chebyshev_t`, so that multiple allocations are
prevented.