Allocate `positions` within `pcomp_find_chebyshev_mask`
This commit partially reverts 67ccd59722, as the `positions` array is now
allocated locally within `pcomp_find_chebyshev_mask`. The reason is that when
the code runs on multiple processors, `positions` might be accessed and written
by different threads at the same time.

A few checks show that allocating `positions` in the heap does not impact
significantly the performance with respect to a global allocation (as it was
the case before this commit) or calling `alloca`.