Remove OpenMP support from `pcomp_decompress_polycomp`
It is too difficult to implement a viable way to implement a OpenMP loop in
this function. Previous commits introduced race conditions, due to the fact
that they called `pcomp_init_chebyshev` while in an OpenMP thread: this is not
allowed, as specified in the FFTW documentation
(see http://www.fftw.org/doc/Thread-safety.html).

The solution would be to allocate everything before entering the decompression
loop. But this is not as easy as it sounds:

1. We cannot be sure that every chunk will have the same size, so we should
first see how many sizes need to be allocated and then create a new structure
for each of them.

2. We could assume that all the chunks but the last one have the same size,
like it has been done in `pcomp_compress_polycomp`.  This is probably the most
viable solution.